Add configuration to compile as C++14.
Display the graph weight labels smaller, and in a neat fashion, so that the text hopefully fits in the allocated space.
Change remove DB button on dashboard page to ask confirmation.
Add method to find the minimum point count for all exercises in the DB, to know when to show the graph component.
Display a placeholder text when there are not enough points to show the graph.

+++ LiftLog/pages/DashboardPage.qml
@@ -92,11 +92,22 @@ BasicPage {
             Layout.fillWidth: true
             Layout.fillHeight: true
             onClicked: {
-                dbManager.deleteDBFileAndReInit()
-                pageStack.showWelcomePage()
+                showModalPopup()
+                modalPopup.prepare("deletedb")
             }
             onStateChanged: buttonStageChange(state)
         }
+
+        Connections {
+            target: modalPopup
+            onAcceptClicked: {
+                if (modalPopup.operation === modalPopup.popupDeleteDBOperation) {
+                    dbManager.deleteDBFileAndReInit()
+                    pageStack.showWelcomePage()
+                }
+            }
+        }
+
 
         //            Label {
         //                text: qsTr("Welcome back, %1").arg(appState.currentUser.name)
